Make CKeditor work with the back button
Using the back button caused the html-filled output of the submission
to appear in the textarea, then be encoded _again_ by the
ckeditor-creation code.  We work around this by inserting a
<input type="text" style="display:none"> element into the form, which
we update when we upgrade the field.  This serves dual purposes: it
notifies the perl side of the content type of the field, but it also
serves as a storage of if the text->html encoding has already happened
in javascript.

This needs to be a type="text" because, unlike text="hidden", values
in those fields are restored to their submit-time values when the back
button is pressed.  Thus, the only way to back this system safe from
the back button is to store that information in a non-hidden text
area, and hide it using CSS.