Enforce loading of Module::Refresh, whenever and however DevelMode gets set
t/web/squish.t uses the RT::Test infrastructure to set DevelMode halfway
through a test file, which walks around the standard codepaths in
RT::LoadConfig where we tried to require Module::Refresh.  Move the
require to just before the module is used, to ensure that it is loaded
if necessary.  While this imposes a slight run-time penalty to each
request, it can be safely assumed that if one is running in DevelMode,
extra milliseconds are not a priority.