Pick theme colors directly from the logo
Currently supported by any browser that supports the <canvas> element.
Browsers which don't simply don't see the feature.

Alpha channel information is currently discarded, which can lead to
unexpected results when picking colors that make use of alpha.  Ideally
the color wheel is modified to support alpha, and we just start setting
it in the CSS with rgba().